Changed the playlist time based on SongUI
